+++ apps/desktop/src/app/vault/view-custom-fields.component.ts
@@ -1,7 +1,7 @@
 import { Component } from "@angular/core";
 
 import { ViewCustomFieldsComponent as BaseViewCustomFieldsComponent } from "@bitwarden/angular/components/view-custom-fields.component";
-import { EventService } from "@bitwarden/common/abstractions/event.service";
+import { EventService } from "@bitwarden/common/abstractions/event/event.service";
 
 @Component({
   selector: "app-vault-view-custom-fields",
