+++ apps/web/src/app/core/init.service.ts
@@ -8,14 +8,14 @@ import {
   EnvironmentService as EnvironmentServiceAbstraction,
   Urls,
 } from "@bitwarden/common/abstractions/environment.service";
-import { EventService as EventLoggingServiceAbstraction } from "@bitwarden/common/abstractions/event.service";
+import { EventService as EventLoggingServiceAbstraction } from "@bitwarden/common/abstractions/event/event.service";
 import { I18nService as I18nServiceAbstraction } from "@bitwarden/common/abstractions/i18n.service";
 import { NotificationsService as NotificationsServiceAbstraction } from "@bitwarden/common/abstractions/notifications.service";
 import { StateService as StateServiceAbstraction } from "@bitwarden/common/abstractions/state.service";
 import { TwoFactorService as TwoFactorServiceAbstraction } from "@bitwarden/common/abstractions/twoFactor.service";
 import { VaultTimeoutService as VaultTimeoutServiceAbstraction } from "@bitwarden/common/abstractions/vaultTimeout/vaultTimeout.service";
 import { ContainerService } from "@bitwarden/common/services/container.service";
-import { EventService as EventLoggingService } from "@bitwarden/common/services/event.service";
+import { EventService as EventLoggingService } from "@bitwarden/common/services/event/event.service";
 import { VaultTimeoutService as VaultTimeoutService } from "@bitwarden/common/services/vaultTimeout/vaultTimeout.service";
 
 import { I18nService } from "./i18n.service";
