+++ libs/angular/src/components/view-custom-fields.component.ts
@@ -1,6 +1,6 @@
 import { Directive, Input } from "@angular/core";
 
-import { EventService } from "@bitwarden/common/abstractions/event.service";
+import { EventService } from "@bitwarden/common/abstractions/event/event.service";
 import { EventType } from "@bitwarden/common/enums/eventType";
 import { FieldType } from "@bitwarden/common/enums/fieldType";
 import { CipherView } from "@bitwarden/common/models/view/cipher.view";
