+++ libs/common/src/services/event/event.service.ts
@@ -1,12 +1,12 @@
-import { ApiService } from "../abstractions/api.service";
-import { CipherService } from "../abstractions/cipher.service";
-import { EventService as EventServiceAbstraction } from "../abstractions/event.service";
-import { LogService } from "../abstractions/log.service";
-import { OrganizationService } from "../abstractions/organization/organization.service.abstraction";
-import { StateService } from "../abstractions/state.service";
-import { EventType } from "../enums/eventType";
-import { EventData } from "../models/data/event.data";
-import { EventRequest } from "../models/request/event.request";
+import { ApiService } from "../../abstractions/api.service";
+import { CipherService } from "../../abstractions/cipher.service";
+import { EventService as EventServiceAbstraction } from "../../abstractions/event/event.service";
+import { LogService } from "../../abstractions/log.service";
+import { OrganizationService } from "../../abstractions/organization/organization.service.abstraction";
+import { StateService } from "../../abstractions/state.service";
+import { EventType } from "../../enums/eventType";
+import { EventData } from "../../models/data/event.data";
+import { EventRequest } from "../../models/request/event.request";
 
 export class EventService implements EventServiceAbstraction {
   private inited = false;
